added reload and shoot sounds
added some basic sounds for the reload and shoot methods of the Gun scirpt and set according variables for the Pistol and the shotgun.
Also added a mild camera shake to the Camera. I tried to modify it so it is stronger when sprinting, but that didn't work out, as the camera would snap back whenever one starts/stops sprinting