death animation as well as hit animation and bugfix for patrol state animation
enemy gameobject now only gets destroyed after the animation is done. This could be confusing, as the enemy is somewhat still looking like a target to shoot and it still gives hitmarkers.
Also added animations for when the enemy gets shot. However, because of the previous change where the enmy can only shoot when he is in the "holding weapon" state, he can't shoot back for half a sec after being shot. This could lead to a situation where constantly hitting the enemy never gives him a chance to hit back.
Also fixed a bug with the animations in the Patrol state,