sound option working
the slider should now actually change the sounds. The slider volume change is a bit un-equal. So you probably want the slider in the most right third or otherwise you hear almost nothing
I added seperate mixers for Music, but it seems probably not that important anyways and would  maybe not work perfectly.
Also added a helper class that finds all Audio sources in each scene