added sfx for enemies
I didn't find any death sounds, so the zombie has a scream as death sound and the alien got nothing. The other effects aren't that great, either.
And I don't think I managed to have the sounds play at the position of the enemies, but rather they play like music from all sides of the player, so to say